osx: make sure that dialog windows are menu-less
on linux, when modal dialog is opened application menu shortcuts
are not active and user can type characters even if they are
menu shortcuts.
However on OSX it's not so and if shortcut is a normal character
that user tries to type into input field of dialog window,
keystroke will be intercepted by menu and result in connected action.

To fix issue assign empty menu when creating toplevel windows so that
menu-less dialogs wouldn't have any menu/accelerators by default to
avoid issue.
Windows that have a dedicater menu install it after gtkut_window_new(),
so initial empty menu is replaced explicitly for them later
(mainwindow, compose, addressbook)

Signed-off-by: Igor Mammedov <imammedo@redhat.com>